chore: use root Node min version for new packages
